Rename pages to appropriate naming
